Handle shape distance traveled as double as long as possible to avoid rounding errors. (We are experiencing this when shape_dist_traveled for last stop time is 0.001 greater than total shape dist in some cases)

+++ mobi.chouette.exchange.gtfs/src/main/java/mobi/chouette/exchange/gtfs/exporter/producer/GtfsShapeProducer.java
@@ -44,7 +44,7 @@ public class GtfsShapeProducer extends AbstractProducer
 		   return false;
 	   int shapePtSequence = 0;
 	   int startIndex = 0;
-	   float distance = (float) 0.0;
+	   double distance = 0.0;
 
 	   for (RouteSection rs : neptuneObject.getRouteSections() ) {
 		   shape.setShapeId(toGtfsId(neptuneObject.getObjectId(), prefix, keepOriginalId));
@@ -70,10 +70,10 @@ public class GtfsShapeProducer extends AbstractProducer
 			   shape.setShapePtSequence(shapePtSequence++);
 			   if (i > 0)
 			   {
-				   distance += (float) computeDistance(prev,cs[i]);
+				   distance +=  computeDistance(prev,cs[i]);
 				   prev = cs[i];
 			   }
-			   shape.setShapeDistTraveled(Float.valueOf(distance));
+			   shape.setShapeDistTraveled(Float.valueOf((float)distance));
 			   try {
 				   getExporter().getShapeExporter().export(shape);
 			   } catch (Exception e) {
